server refactored

+++ server/tests/connection.js
@@ -1,16 +1,19 @@
-const mongoose = require("mongoose");
+const mongoose = require('mongoose');
 
 module.exports = {
-    openConnect (done) {
-      mongoose.connect("mongodb://localhost:27017/JestDB",
-        { useNewUrlParser: true, useUnifiedTopology: true },
-        () => done());
+    async openConnect (done) {
+      await mongoose.connection.close()
+      await mongoose.connect('mongodb://localhost:27017/JestDB',
+      { useNewUrlParser: true, useUnifiedTopology: true },
+      () => done());
     },
     closeConnection (done) {
+      if (!mongoose.connection.db) {
+        done()
+        return
+      }
       mongoose.connection.db.dropDatabase(() => {
         mongoose.connection.close(() => done())
       });      
     }
 }
-
-
